pyproject.toml
[project]
name = "document-rag-system"
version = "0.1.0"
description = ""
authors = [
    {name = "Azam Afridi",email = "azam.afridi@tkxel.io"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi (>=0.115.13,<0.116.0)",
    "python-multipart (>=0.0.9,<1.0.0)",
    "uvicorn[standard] (>=0.34.3,<0.35.0)",
    "pydantic-settings (>=2.10.0,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "langchain-openai (>=0.3.25,<0.4.0)",
    "langchain-postgres (>=0.0.15,<0.0.16)",
    "langchain (>=0.3.26,<0.4.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "docx2txt (>=0.9,<0.10)",
    "langchain-community (>=0.3.26,<0.4.0)",
    "pymupdf (>=1.26.1,<2.0.0)",
    "google-api-python-client (>=2.174.0,<3.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "google-auth-oauthlib (>=1.2.2,<2.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "pymupdf4llm (>=0.0.25,<0.0.26)",
    "boto3 (>=1.39.0,<2.0.0)",
    "psycopg-binary (>=3.2.9,<4.0.0)",
    "streamlit (>=1.36.0,<2.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "google-auth (>=2.40.3,<3.0.0)",
    "aioredis (>=2.0.1,<3.0.0)",
    "flower (>=2.0.1,<3.0.0)",
    "twilio (>=9.6.5,<10.0.0)"
]

[project.optional-dependencies]
evaluation = [
    "ragas (>=0.2.15,<0.3.0)",
    "datasets (>=3.6.0,<4.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
